% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{pd_text}
\alias{pd_text}
\alias{pd_token}
\alias{pd_start_line}
\alias{pd_end_line}
\alias{pd_filename}
\alias{pd_start_col}
\alias{pd_end_col}
\title{Accessor functions}
\usage{
pd_text(id, pd)
}
\arguments{
\item{id}{the ID of the expression}

\item{pd}{the parse data.}
}
\description{
This collection of function can be used to easily access elements of
the parse data information.
}
