% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{all_grouping_ids}
\alias{all_grouping_ids}
\title{get the grouping ids}
\usage{
all_grouping_ids(pd = get("pd", parent.frame()))
}
\arguments{
\item{pd}{The \code{\link{parse-data}} information}
}
\value{
an integer vector of ids.
}
\description{
get the ids that represent the grouping nodes.
}
