% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iff_blocks.R
\name{all_iff_ids}
\alias{all_iff_ids}
\title{Identify all \code{if(FALSE)} blocks ids.}
\usage{
all_iff_ids(pd, root.only = TRUE, ignore.groups = FALSE, ...)
}
\arguments{
\item{pd}{The \code{\link{parse-data}} information}

\item{root.only}{only root blocks(`TRUE`) or all block (`FALSE`)}

\item{ignore.groups}{Ignore code grouping}

\item{...}{passed to \code{\link{pd_is_iff_block}}}
}
\value{
an integer vector of all ids identifying
  \code{\link[=iff-blocks]{if(FALSE)}}\link[=iff-blocks]{ blocks}.
}
\description{
Retreives all the ids from pd that identify
  \code{\link[=iff-blocks]{if(FALSE)}} blocks.
  See \code{\link{pd_is_root}} for details on \code{root.only}.
  See \code{\link{pd_is_grouping}} for details on groups affected by
  \code{ignore.groups}.
}
