% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{all_root_nodes}
\alias{all_root_nodes}
\title{Find all root node from parse data}
\usage{
all_root_nodes(pd, include.groups = TRUE)
}
\arguments{
\item{pd}{The \code{\link{parse-data}} information}

\item{include.groups}{descend into grouped code \code{\{\}}?}
}
\value{
\code{\link{parse-data}} with for the root nodes.
}
\description{
A root node in a file is a standalone expression, such as in
source file a function definition.
when discussing a subset it is any expression that does not have
a parent in the subset.
}
