% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{clean_tag_comments}
\alias{clean_tag_comments}
\title{clean tag comments}
\usage{
clean_tag_comments(x, tag)
}
\arguments{
\item{x}{text to strip from}

\item{tag}{tag(s) to remove}
}
\description{
replaces '\code{#@tag}' with '\code{#! @tag}'
}
