% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_blocks.R
\name{extract_test_block}
\alias{extract_test_block}
\title{Extract testing blocks from the parse-data.}
\usage{
extract_test_block(
  id = all_tagged_iff_block_ids(pd, .testing.tags),
  pd = get("pd", parent.frame())
)
}
\arguments{
\item{id}{iff block id, not the content}

\item{pd}{a \link{parse-data} object.}
}
\value{
a character vector with the lines for the specific test(s)
with the name of the test included as an attribute.
}
\description{
Extract the content of a testing block as a character vector of lines.
The name, which is attached as an attribute is taken from the info
string or inferred by location, see Details.
}
