% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_blocks.R
\name{extract_test_blocks}
\alias{extract_test_blocks}
\title{extract tests from a file.}
\usage{
extract_test_blocks(file)
}
\arguments{
\item{file}{the file to retrieve tests from.}
}
\description{
Convenience function for extracting all tests from a file.
This parses the file and passes the work to
\code{\link{extract_test_block}}.
}
