% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parse_data.R
\name{get_parse_data}
\alias{get_parse_data}
\title{Get cleaned parse data}
\usage{
get_parse_data(x, ...)
}
\arguments{
\item{x}{an object to get parse-data from.}

\item{...}{passed on}
}
\description{
A customized version of `<getParseData>` that will return 
a cleaned up version of the parse data for a variety of objects.
This version also fails less often, even reparsing text when 
needed.
}
