% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{has_tag}
\alias{has_tag}
\title{Check if there is a documentation `@` tag.}
\usage{
has_tag(pd, tag, id = pd$id, ...)
}
\arguments{
\item{tag}{tag to test for}

\item{...}{options passed on}
}
\description{
Check if a node of \code{parse-data} identified by \code{id}
is both a comment and contains a documentation tag itentifed by
the `@` symbol.
}
