% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{n_children}
\alias{n_children}
\title{Count the number of children}
\usage{
n_children(id = pd$id, pd = get("pd", parent.frame()))
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}
}
\description{
Count the number of children
}
