% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iff_blocks.R
\name{pd_all_tagged_iff_block_ids}
\alias{pd_all_tagged_iff_block_ids}
\title{Find all tagged \code{if(FALSE)} blocks.}
\usage{
pd_all_tagged_iff_block_ids(pd, tag, doc.only = TRUE)
}
\arguments{
\item{pd}{The \code{\link{parse-data}} information}

\item{tag}{The tag to consider.}

\item{doc.only}{Should comments be restricted to documentation style
comments only?}
}
\value{
an integer vector giving the ids in \code{pd} that identify
\code{\link[=iff-blocks]{if(FALSE)}}\link[=iff-blocks]{ blocks}
that are also tagged with \code{tag}.
}
\description{
Retrieves all ids identifying \code{\link[=iff-blocks]{if(FALSE)}}
blocks that are also tagged with \code{tag}.
See \code{\link{pd_is_tagged_iff_block}} for details.
}
\seealso{
\code{\link{pd_is_iff_block}}, \code{\link{pd_is_tagged_iff_block}},
\code{\link{pd_has_tag}}
}
