% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{pd_get_ancestor_ids}
\alias{pd_get_ancestor_ids}
\title{Get the ancestors of id in pd.}
\usage{
pd_get_ancestor_ids(id, pd, nancestors = Inf, aggregate = TRUE,
  include.self = TRUE, only.present = FALSE, last = 0L, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{nancestors}{Number of generations to go back}

\item{aggregate}{All (T) or only final (F).}

\item{include.self}{should `id` be included in list of ancestors
if `aggregate` is true.}

\item{only.present}{should the list be restricted to only those
node that are present?  Most revelvent for
when parent is zero.}

\item{last}{the last acceptable parent.}

\item{.check}{Perform checks for input validation?}
}
\description{
Get the ancestors for \code{id}.  See argument descriptions for details.
}
