% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_assign.R
\name{pd_get_assign_value_id}
\alias{pd_get_assign_value_id}
\title{Get the id for the value portion of an assignment}
\usage{
pd_get_assign_value_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
an id integer.
}
\description{
Gives the id of the value portion of the assignment, while correctly
accounting for the direction of the arrow.
}
