% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_assign.R
\name{pd_get_assign_variable_id}
\alias{pd_get_assign_variable_id}
\title{Get the variable of an assignment}
\usage{
pd_get_assign_variable_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
an id integer, typically pointing to an 'expr' node in pd.
}
\description{
Gets the id for the variable portion of an assignment expression.
This accounts for the direction of the assignment arrow.
}
