% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_call.R
\name{pd_get_call_arg_ids}
\alias{pd_get_call_arg_ids}
\title{get ids of the arguments of a call.}
\usage{
pd_get_call_arg_ids(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
a named list where each element is the id for the 'expr' element of the argument.
}
\description{
Retrieves the ids of the arguments of a call as an integer vector.
}
\note{
Are there cases other than alist that could result in a two
id argument?
}
