% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_call.R
\name{pd_get_call_symbol_id}
\alias{pd_get_call_symbol_id}
\title{Get the symbol of the function being called.}
\usage{
pd_get_call_symbol_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Gets the id of the symbol of the call.
   That is the name of the function being called.
}
