% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{pd_get_children_ids}
\alias{pd_get_children_ids}
\alias{get_children_pd}
\title{Get all nodes that are children of `id`.}
\usage{
pd_get_children_ids(id, pd, ngenerations = 1, include.self = FALSE,
  all.generations = TRUE, .check = TRUE)

get_children_pd(id, pd, ..., .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{ngenerations}{Number of levels to descend.}

\item{include.self}{Should the root node (\code{id}) be included?}

\item{all.generations}{Should all generations(TRUE) or only the
the final (FALSE) generation be returned?}

\item{.check}{Perform checks for input validation?}

\item{...}{arguments passed on.}
}
\description{
Get all ids in `pd` that are children of \code{id}.
  i.e. lower in the heirarchy or with id as a parent.
  If \code{ngenerations} is greater than 1 and \code{all.generations}
  is \code{TRUE}, all descendents are aggregated and returned.
}
