% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_classes.R
\name{pd_get_closest_call_id}
\alias{pd_get_closest_call_id}
\title{Get the closest call ID.}
\usage{
pd_get_closest_call_id(id, pd, calls = NULL, .check = TRUE)
}
\arguments{
\item{calls}{optional calls to limit consideration to.}
}
\description{
Get the id of the call that is closest to the \code{id} given.
Closest is defined as the innermost call that contains the \code{id}.
}
