% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pd_get_comment_tag_content}
\alias{pd_get_comment_tag_content}
\title{Get the content of a tag}
\usage{
pd_get_comment_tag_content(id, pd, tag, all.contiguous = FALSE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{tag}{tag(s) to test for}

\item{all.contiguous}{if TRUE get all comments connected to this element.}
}
\description{
Get the content of a tag
}
