% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{pd_get_family}
\alias{pd_get_family}
\title{Get family of nodes.}
\usage{
pd_get_family(id, pd, include.self = TRUE, ngenerations = Inf, ...,
  include.doc.comments = TRUE, include.regular.comments = FALSE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{include.self}{Should the root node (\code{id}) be included?}

\item{ngenerations}{Number of levels to descend.}

\item{...}{currently ignored.}

\item{include.doc.comments}{include associated documentation comments.}

\item{include.regular.comments}{include associated regular comments.}
}
\value{
a subset of the \code{\link{parse-data}} \code{pd}.
}
\description{
Subset the \code{pd} to the family of \code{id}.
}
