% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{pd_get_firstborn_id}
\alias{pd_get_firstborn_id}
\title{get the firstborn child.}
\usage{
pd_get_firstborn_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Get the id of the firstborn child of id.
  Without the "_id" is a wrapper for giving the nodes.
}
