% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_function.R
\name{pd_get_function_arg_variable_ids}
\alias{pd_get_function_arg_variable_ids}
\alias{pd_get_function_arg_variable_text}
\alias{pd_is_function_arg}
\alias{pd_get_function_arg_associated_comment_ids}
\title{Retrieve the variable for a function argument}
\usage{
pd_get_function_arg_variable_ids(id, pd, .check = TRUE)

pd_get_function_arg_variable_text(id, pd, .check = TRUE)

pd_is_function_arg(id, pd, .check = TRUE)

pd_get_function_arg_associated_comment_ids(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Retrieve the variable for a function argument
}
