% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_if.R
\name{pd_get_if_alternate_id}
\alias{pd_get_if_alternate_id}
\title{Get the alternate branch of if statement}
\usage{
pd_get_if_alternate_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
an id integer.
}
\description{
Gets the id of the alternate branch, i.e. the else branch.
}
