% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_if.R
\name{pd_get_if_branch_id}
\alias{pd_get_if_branch_id}
\title{Get branch of if statment.}
\usage{
pd_get_if_branch_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
an id integer.
}
\description{
Returns the id of the body of the branch executed if the predicate
evaluates to true.
}
