% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_if.R
\name{pd_get_if_predicate_id}
\alias{pd_get_if_predicate_id}
\title{Get if predicate id}
\usage{
pd_get_if_predicate_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Returns the id of the predicate of the if statemment,
  i.e. the conditional statement.
}
