% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{pd_get_parent_id}
\alias{pd_get_parent_id}
\title{Get the parent of the expression identified by `id` in `pd`.}
\usage{
pd_get_parent_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Get the parent of the expression identified by `id` in `pd`.
}
