% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_comments.R
\name{pd_get_relative_comment_associated_ids}
\alias{pd_get_relative_comment_associated_ids}
\title{Associate relative documentation comments}
\usage{
pd_get_relative_comment_associated_ids(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
Returns a vector of the same length as id.  Where the value is
either the id of the associated object or NA if it cannot be
associated.
}
\description{
Relative comment created with \code{\#\<} comment tags document something
designated by the location of the comment.
In general, the comment documents the previous symbol.
A comment will not be associated with any parse id that does not have
the same parent as the comment. For example,
}
\details{
\preformatted{function(x #< a valid comment
        )\{\}
}

would associate \code{a valid comment} with \code{x}, but\preformatted{function(x)\{ #< not a valid comment
           \}
}

would not.
}
