% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{pd_get_sibling_ids}
\alias{pd_get_sibling_ids}
\alias{pd_get_next_sibling_id}
\alias{pd_get_prev_sibling_id}
\title{Identify siblings.}
\usage{
pd_get_sibling_ids(id, pd, .check = TRUE)

pd_get_next_sibling_id(id, pd, .check = TRUE)

pd_get_prev_sibling_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
\subsection{pd_get_sibling_ids}{
  A convenience function for identifying siblings of the given id.
  Siblings are nodes with the same parent.
}

\subsection{pd_get_next_sibling_id}{
  gives the id of the next youngest sibling of the current id.
}

\subsection{pd_get_prev_sibling_id}{
  gives the id of the next older sibling of the current id.
}
}
