% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pd_get_tagged_comment_ids}
\alias{pd_get_tagged_comment_ids}
\title{Get tagged comment ids}
\usage{
pd_get_tagged_comment_ids(pd, tag, doc.only = TRUE)
}
\arguments{
\item{pd}{The \code{\link{parse-data}} information}

\item{tag}{tag(s) to test for}

\item{doc.only}{Restrict to documentation comments only?}
}
\value{
an integer vector of ids.
}
\description{
Finds all ids that are comments and contain the given '@' \code{tag}.
  If doc.only is true(default) then only documentation comments are
  considered, otherwise all comments are examined.
}
