% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_identify.R
\name{pd_identify}
\alias{pd_identify}
\alias{pd_identify.default}
\alias{pd_identify.NULL}
\alias{pd_identify.srcref}
\title{Get the ID for an object}
\usage{
pd_identify(pd, object)

\method{pd_identify}{default}(pd, object)

\method{pd_identify}{NULL}(pd, object)

\method{pd_identify}{srcref}(pd, object)
}
\arguments{
\item{pd}{the parse data.}

\item{object}{an object that originated in pd,
for which to obtain the ID.}
}
\description{
Identify in pd the id for the object given.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method identifies by \code{\link[base:srcref]{base::srcref()}}.

\item \code{NULL}: Passing a NULL object will result in an error.

\item \code{srcref}: Identify by explicit \code{srcref}.
}}

