% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_assign.R
\name{pd_is_assignment}
\alias{pd_is_assignment}
\title{Check if it is an assignment}
\usage{
pd_is_assignment(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
check if the provided parse-data is an assignment expression.
}
