% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_call.R
\name{pd_is_call}
\alias{pd_is_call}
\title{Is a call?}
\usage{
pd_is_call(id, pd, calls = NULL, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{calls}{Acceptable calls, if \code{NULL} (default) all calls allowed.}

\item{.check}{Perform checks for input validation?}
}
\description{
Checks if the \code{id} identifies in \code{pd} a call expression.
}
