% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{pd_is_firstborn}
\alias{pd_is_firstborn}
\title{Test if id is the firstborn.}
\usage{
pd_is_firstborn(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
Test if an expression is the firstborn, ie. oldest or lowest id.
}
