% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_function.R
\name{pd_is_function}
\alias{pd_is_function}
\alias{pd_get_function_body_id}
\alias{pd_get_function_arg_ids}
\title{test if a function}
\usage{
pd_is_function(id, pd, .check = TRUE)

pd_get_function_body_id(id, pd, .check = TRUE)

pd_get_function_arg_ids(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
a logical vector, same length as \code{id}.
}
\description{
Test if the \code{id} points to a function.
}
\section{Functions}{
\itemize{
\item \code{pd_get_function_body_id}: Obtain the body of a function

\item \code{pd_get_function_arg_ids}: Obtain the ids for the arguments of a function
}}

