% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_if.R
\name{pd_is_if}
\alias{pd_is_if}
\title{Is if expression?}
\usage{
pd_is_if(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
a logical vector of same length as id.
}
\description{
Tests if the id(s) represent if expressions.
}
