% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iff_blocks.R
\name{pd_is_iff_block}
\alias{pd_is_iff_block}
\alias{iff-blocks}
\title{test if an expresion ID points to a `if(FALSE)` statement.}
\usage{
pd_is_iff_block(id, pd, allow.short = TRUE, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{allow.short}{Should \code{F} be interpreted as FALSE.}

\item{.check}{Perform checks for input validation?}
}
\value{
A logical vector of same length as id indicating if the id
     represents a \code{if(FALSE)} block.
}
\description{
This function tests if an expression id is the root of an
  \code{if(FALSE)} block, which many users use to deactive code but
  can also be used to support including examples and testing code
  in the same file as the source code.
}
