% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_call.R
\name{pd_is_symbol_call}
\alias{pd_is_symbol_call}
\title{Check if the call is specifically a symbol call}
\usage{
pd_is_symbol_call(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\value{
a logical of the same length as \code{id}
}
\description{
Checks if the \code{id} identifies in \code{pd} specifically a
symbol call expression, That is a call from a symbol.
}
