% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iff_blocks.R
\name{pd_is_tagged_iff}
\alias{pd_is_tagged_iff}
\title{Test tagged \code{if(FALSE)} blocks.}
\usage{
pd_is_tagged_iff(id, tag, pd, doc.only = TRUE, ...)
}
\arguments{
\item{id}{id of the expression of interest}

\item{tag}{tag(s) to test for}

\item{pd}{The \code{\link{parse-data}} information}

\item{doc.only}{Should comments be restricted to documentation style
comments only?}

\item{...}{options passed on}
}
\value{
a logical vector indicating if the \code{id} in \code{pd}
     identifies an \code{\link[=iff-blocks]{if(FALSE)}}
     \link[=iff-blocks]{ block} that also has the tag identified
     in \code{tag}.
}
\description{
This functions tests if an id is: \enumerate{


  \item an \code{if(FALSE)} block.
  \item is a curly braced group of code.
  \item has a comment as the first parsed element.
  \item and that it's a documentation comment if doc.only is true.
  \item and finally that the comment contains the identified \code{tag(s)}.
  }
}
\seealso{
\code{\link{pd_is_iff_block}}, \code{\link{pd_has_tag}}
}
