% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_make_is_in.R
\name{pd_make_is_in_call}
\alias{pd_make_is_in_call}
\alias{pd_make_is_call}
\title{Create a function to test if an id is contained in a type}
\usage{
pd_make_is_in_call(calls, .is = pd_make_is_call(calls))

pd_make_is_call(calls)
}
\arguments{
\item{calls}{The tokens to test against}

\item{.is}{A function to test if a specific id is a valid}
}
\description{
Create a function to test if an id is contained in a type
}
