% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{strip_doc_comment_leads}
\alias{strip_doc_comment_leads}
\title{Remove the characters identifying a documentation comment.}
\usage{
strip_doc_comment_leads(comment, rm.space = TRUE)
}
\arguments{
\item{comment}{The text of the comments or parse data.}

\item{rm.space}{should the space at the beginning of the line be removed.}
}
\description{
Remove the characters identifying a documentation comment as a 
document comment leaving only the relevant text.
}
