% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{strip_tag}
\alias{strip_tag}
\title{Remove a tag that identified a line.}
\usage{
strip_tag(x, tag, ...)
}
\arguments{
\item{x}{text to strip from}

\item{tag}{tag(s) to remove}

\item{...}{passed on options]}
}
\description{
Removes \code{@tag} tags from the text.
Also will remove '\code{#@tag}' replacing with '\code{#!}'.
}
