% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_assign.R
\name{assignments}
\alias{assignments}
\alias{pd_is_assignment}
\alias{pd_get_assign_value_id}
\alias{pd_get_assign_variable_id}
\title{Assignment Node Navigation.}
\usage{
pd_is_assignment(id, pd, .check = TRUE)

pd_get_assign_value_id(id, pd, .check = TRUE)

pd_get_assign_variable_id(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
These function help identify and navigate assignments in parse data.
}
\details{
These functions only deal with assignment operators.
Using \code{\link[base:assign]{base::assign()}} or \code{\link[base:delayedAssign]{base::delayedAssign()}} are considered
calls in terms of parse data.

There are five assignment operators grouped into three categories.
\itemize{
\item Left assignment, the \href{`<-`}{base::assignOps} and \href{`<<-`}{base::assignOps},
\item right assignment, \href{`->`}{base::assignOps} and the rarely used \href{`->>`}{base::assignOps}
\item and the equals assignment \href{`=`}{base::assignOps}.
}
}
\section{Functions}{
\itemize{
\item \code{pd_is_assignment}: Check if the node is an assignment expression.

\item \code{pd_get_assign_value_id}: Get the id for the value portion of an assignment.

\item \code{pd_get_assign_variable_id}: Get the variable of an assignment.
}}

