% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{pd_is_grouping}
\alias{pd_is_grouping}
\title{test if an id is a grouping element}
\usage{
pd_is_grouping(id, pd, .check = TRUE)
}
\arguments{
\item{id}{id of the expression of interest}

\item{pd}{The \code{\link{parse-data}} information}

\item{.check}{Perform checks for input validation?}
}
\description{
a grouping is defined as a non empty set
started with a "'\{'" token and
and there is no parent or the parent is also a grouping.
}
